Changes for 1.2.2: Update versions to 1.2.2. Modify signature of makeVideoJustification. Remove VideoJustificationChannelBoth. Make aida:channel required. Update tests

+++ python/tests/py_validator.sh
@@ -13,12 +13,20 @@ DIR_PATH=$test_directory python3 examples.py
 
 # validate test file directory
 # if any files are invalid, the script will end (see above)
-docker run --rm -it \
-       --user $(id -u):$(id -g) \
-       -v $(pwd)/$test_directory:/v \
-       --entrypoint /opt/aif-validator/java/target/appassembler/bin/validateAIF \
-       nextcenturycorp/aif_validator:latest \
-       -o --ont /opt/aif-validator/java/src/main/resources/com/ncc/aif/ontologies/LDCOntology -d /v
+if [ "$1" = "--local" ]; then
+    # save location of LDCOntology (validateAIF requires absolute path)
+    pushd ../../java/src/main/resources/com/ncc/aif/ontologies
+    ontology=`pwd`/LDCOntology
+    popd
+    ../../java/target/appassembler/bin/validateAIF --ont $ontology -o -d $test_directory
+else
+    docker run --rm -it \
+        --user $(id -u):$(id -g) \
+        -v $(pwd)/$test_directory:/v \
+        --entrypoint /opt/aif-validator/java/target/appassembler/bin/validateAIF \
+        nextcenturycorp/aif_validator:latest \
+        -o --ont /opt/aif-validator/java/src/main/resources/com/ncc/aif/ontologies/LDCOntology -d /v
+fi
 
 # if all files are valid, delete the test file directory
 rm -r $test_directory
